+++ pyramid_openapi3/tests/test_views.py
@@ -13,6 +13,7 @@ from pyramid.router import Router
 from pyramid.testing import DummyRequest
 from pyramid.testing import testConfig
 from pyramid_openapi3.exceptions import RequestValidationError
+from pyramid_openapi3 import check_all_routes
 
 import os
 import pytest
@@ -37,6 +38,19 @@ MINIMAL_DOCUMENT = b"""
           responses:
             200:
               description: A foo
+"""
+
+ALTERNATE_DOCUMENT = b"""
+    openapi: "3.0.0"
+    info:
+      version: "1.0.0"
+      title: Bar API
+    paths:
+      /bar:
+        get:
+          responses:
+            200:
+              description: A bar
 """
 
 SPLIT_DOCUMENT = b"""
@@ -263,6 +277,60 @@ def test_add_explorer_view() -> None:
         assert b"<title>Swagger UI</title>" in response.body
 
 
+def test_add_multiple_explorer_views() -> None:
+    """Test registration of multiple viewa serving different Swagger UI."""
+    with testConfig() as config:
+        config.include("pyramid_openapi3")
+
+        with tempfile.NamedTemporaryFile() as document:
+            document.write(MINIMAL_DOCUMENT)
+            document.seek(0)
+
+            config.pyramid_openapi3_spec(
+                document.name, route="/foo.yaml", route_name="foo_api_spec",
+                apiname='foo_api'
+            )
+            config.pyramid_openapi3_add_explorer(
+                route=f'/foo_api/v1/',
+                route_name=f'foo_api_explorer',
+                apiname='foo_api',
+            )
+
+        with tempfile.NamedTemporaryFile() as document:
+            document.write(ALTERNATE_DOCUMENT)
+            document.seek(0)
+
+            config.pyramid_openapi3_spec(
+                document.name, route="/bar.yaml", route_name="bar_api_spec",
+                apiname='bar_api'
+            )
+            config.pyramid_openapi3_add_explorer(
+                route=f'/bar_api/v1/',
+                route_name=f'bar_api_explorer',
+                apiname='bar_api',
+            )
+
+        request = config.registry.queryUtility(
+            IRouteRequest, name="foo_api_explorer"
+        )
+        view = config.registry.adapters.registered(
+            (IViewClassifier, request, Interface), IView, name=""
+        )
+        response = view(request=DummyRequest(config=config), context=None)
+        assert b"<title>Swagger UI</title>" in response.body
+        assert b'http://example.com/foo.yaml' in response.body
+
+        request = config.registry.queryUtility(
+            IRouteRequest, name="bar_api_explorer"
+        )
+        view = config.registry.adapters.registered(
+            (IViewClassifier, request, Interface), IView, name=""
+        )
+        response = view(request=DummyRequest(config=config), context=None)
+        assert b"<title>Swagger UI</title>" in response.body
+        assert b'http://example.com/bar.yaml' in response.body
+
+
 def test_explorer_view_missing_spec() -> None:
     """Test graceful failure if explorer view is not registered."""
     with testConfig() as config:
@@ -311,6 +379,67 @@ def test_openapi_view() -> None:
         )
         request = DummyRequest(config=config, content_type="text/html")
         request.matched_route = DummyRoute(name="foo", pattern="/foo")
+        context = None
+        response = view(context, request)
+
+        assert response.json == "bar"
+
+
+def test_multiple_openapi_views() -> None:
+    """Test registration multiple openapi views."""
+    with testConfig() as config:
+        config.include("pyramid_openapi3")
+
+        with tempfile.NamedTemporaryFile() as document:
+            document.write(MINIMAL_DOCUMENT)
+            document.seek(0)
+
+            config.pyramid_openapi3_spec(
+                document.name, route="/foo.yaml", route_name="foo_api_spec",
+                apiname='foo'
+            )
+
+        with tempfile.NamedTemporaryFile() as document:
+            document.write(ALTERNATE_DOCUMENT)
+            document.seek(0)
+
+            config.pyramid_openapi3_spec(
+                document.name, route="/bar.yaml", route_name="bar_api_spec",
+                apiname='bar'
+            )
+
+        config.add_route("foo", "/foo")
+        view_func = lambda *arg: "foo"  # noqa: E731
+        config.add_view(openapi=True, renderer="json", view=view_func, route_name="foo")
+
+        config.add_route("bar", "/bar")
+        view_func = lambda *arg: "bar"  # noqa: E731
+        config.add_view(openapi=True, renderer="json", view=view_func, route_name="bar")
+
+        # Simulate, that `check_all_routes` was called
+        settings = config.registry.settings
+        settings.setdefault("pyramid_openapi3", {})
+        settings['pyramid_openapi3'].setdefault('routes', {})
+        settings["pyramid_openapi3"]["routes"]['foo'] = 'foo'
+        settings["pyramid_openapi3"]["routes"]['bar'] = 'bar'
+
+        request_interface = config.registry.queryUtility(IRouteRequest, name="foo")
+        view = config.registry.adapters.registered(
+            (IViewClassifier, request_interface, Interface), IView, name=""
+        )
+        request = DummyRequest(config=config, content_type="text/html")
+        request.matched_route = DummyRoute(name="foo", pattern="/foo")
+        context = None
+        response = view(context, request)
+
+        assert response.json == "foo"
+
+        request_interface = config.registry.queryUtility(IRouteRequest, name="bar")
+        view = config.registry.adapters.registered(
+            (IViewClassifier, request_interface, Interface), IView, name=""
+        )
+        request = DummyRequest(config=config, content_type="text/html")
+        request.matched_route = DummyRoute(name="bar", pattern="/bar")
         context = None
         response = view(context, request)