+++ pyramid_openapi3/tween.py
@@ -32,7 +32,11 @@ def response_tween_factory(handler, registry) -> t.Callable[[Request], Response]
             # validate response
             openapi_request = PyramidOpenAPIRequestFactory.create(request)
             openapi_response = PyramidOpenAPIResponseFactory.create(response)
-            settings = request.registry.settings["pyramid_openapi3"]
+            settings_key = 'pyramid_openapi3'
+            gsettings = settings = request.registry.settings[settings_key]
+            if 'routes' in gsettings:
+                settings_key = gsettings['routes'][request.matched_route.name]
+                settings = request.registry.settings[settings_key]
             result = settings["response_validator"].validate(
                 request=openapi_request, response=openapi_response
             )
@@ -50,7 +54,7 @@ def response_tween_factory(handler, registry) -> t.Callable[[Request], Response]
                             )
                         ),
                         None,
-                        registry.settings["pyramid_openapi3"]["filepath"],
+                        registry.settings[settings_key]["filepath"],
                         0,
                     )
                 raise ResponseValidationError(response=response, errors=result.errors)